# PERF: pd.concat EA-backed indexes and sort=True