--- a/asv_bench/benchmarks/join_merge.py
+++ b/asv_bench/benchmarks/join_merge.py
@@ -97,13 +97,13 @@ class ConcatIndexDtype:
 
     params = (
         ["datetime64[ns]", "int64", "Int64", "string[python]", "string[pyarrow]"],
+        ["monotonic", "non_monotonic", "has_na"],
         [0, 1],
         [True, False],
-        [True, False],
     )
-    param_names = ["dtype", "axis", "sort", "is_monotonic"]
+    param_names = ["dtype", "structure", "axis", "sort"]
 
-    def setup(self, dtype, axis, sort, is_monotonic):
+    def setup(self, dtype, structure, axis, sort):
         N = 10_000
         if dtype == "datetime64[ns]":
             vals = date_range("1970-01-01", periods=N)
@@ -115,14 +115,21 @@ class ConcatIndexDtype:
             raise NotImplementedError
 
         idx = Index(vals, dtype=dtype)
-        if is_monotonic:
+
+        if structure == "monotonic":
             idx = idx.sort_values()
-        else:
+        elif structure == "non_monotonic":
             idx = idx[::-1]
+        elif structure == "has_na":
+            if not idx._can_hold_na:
+                raise NotImplementedError
+            idx = Index([None], dtype=dtype).append(idx)
+        else:
+            raise NotImplementedError
 
-        self.series = [Series(i, idx[i:]) for i in range(5)]
+        self.series = [Series(i, idx[:-i]) for i in range(1, 6)]
 
-    def time_concat_series(self, dtype, axis, sort, is_monotonic):
+    def time_concat_series(self, dtype, structure, axis, sort):
         concat(self.series, axis=axis, sort=sort)